`refactor` Se crea función para reutilizar al descargar archivo
