+++ src/app/containers/layout/layout.component.html
@@ -50,7 +50,7 @@
         </div>
       </ng-container>
       <ng-container class="mt-auto" *ngIf="userLogged?.position?.nom_rol === Nom_rol.ADMINISTRADOR">
-        <div class="mt-auto">
+        <div class="mt-2">
           <small class="text-divider">Herramientas</small>
           <ul class="p-0 my-2 list-none">
               <li *ngFor="let option of menuTools">
